+++ src/new-components/password-field/password-field.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import { useForkRef } from 'reakit-utils'
+import { useForkRef } from 'ariakit-utils'
 import { Tooltip } from '../../components/tooltip'
 import { BaseField } from '../base-field'
 import { Box } from '../box'