+++ CHANGELOG.md
@@ -6,6 +6,8 @@ Reactist follows [semantic versioning](https://semver.org/) and doesn't introduc
 
 -   [Breaking] Dropped support for React 16
     -   Internally, Reactist will now use React 17 and TypeScript v4.6. Targets for `@babel/preset-env` have also been set, reducing the total build size by about 8-10%.
+-   [Breaking] Replace Reakit with Ariakit
+    -   This includes a breaking change for `Tabs`, where if the `selectedId` prop is provided, it becomes a controlled component and will no longer switch tabs on its own. To retain the old behaviour, `defaultSelectedId` can be used instead.
 
 # v11.6.0